Correct the LCDY register update
According to the development manual of Game Boy, the LCDY
indicates the vertical line to which the present data is transferred
to the LCD Driver. It should range from 0 through 153, so we should
reduce LCDY modulo 154.

Also, the function `update_line` should be called after updating the LCDY.
If we update LCDY before `update_line`, the emulator may scroll the wrong
scanline.