Implement fullscreen support (#23)
Fullscreen can be toggled with Alt-Enter. In addition, this patch
uses SDL scancodes instead of keycodes, which fixes issues with
keyboard layouts that have Z and Y swapped.

SDL wiki:
> SDL uses two different types of values to represent keys on the keyboard -
> scancode and keycode.
> SDL_Scancode values are used to represent the physical location of a
> keyboard key on the keyboard.
> SDL_Keycode values are mapped to the current layout of the keyboard and
> correlate to an SDL_Scancode.

* main: remove gratuitous inclusion of execinfo.h

this caused a build error with musl libc, which doesn't provide this
glibc specific header; as none of the functionality it offers is
actually used, it's safe to remove it.